Remove the need for uses_tagged_union deco
At present we support people adding to the list
of Specs after importing the service. This requires a uses_tagged_union decorator on any
BaseModel that references a Spec. We cannot
currently think of a use case for Specs being implemented outside of scanspec.specs,
so removing this requirement means we can ditch the decorator. It also makes the docs (next PR) easier to write:
- Put your specs in src/scanspec/specs.py
- Release a new version
- Update both blueapi and the scanspec service to use that new version