Create CLI name variables in new locals mapping
In 3.13 the behaviour of the locals builtin changed so that
modifications no longer had any affect in certain situations. This meant
the plot CLI subcommand would raise NameError for single letter axis
names as used in the cli tests.

Passing the locals as a new mapping allows the same code to work with
both the new and old locals behaviour.