Updated Specs
Made Product.calculate only accept inner Spec having a .duration

Changed ConstantDuration so that it can receive an empty spec as
argument. If that's the case .calculate we'll return a list with a
single dimension that contains only the duration field.

Updated calculate in ConstantDuration to add duration into the innermost
dimension when fly == False