Made bounds default to false and changed Duration return
Default value of the bounds argument is now `False`. Needed to change
the `frames` method in the `Spec` class so that it would call calculate
with the correct bounds, otherwise it wouls break some tests.

Made it so that the `duration` method in the base class `Spec` returned
None by default and removed the implementations from classes that would
always return None.